+++ lib/utils.js
@@ -1,17 +1,17 @@
-'use strict';
 /**
  * Core Utilities for Sockbot
  * @module utils
  * @license MIT
  */
+'use strict';
+
+const sanitizeHtml = require('sanitize-html');
+
+const storage = new WeakMap();
+
 
 const fs = require('fs');
-const async = require('async');
 const config = require('./config');
-
-const internals = {
-    cooldownTimers: {}
-};
 
 /**
  * Generate a "good enough" Type 4 UUID.
@@ -24,20 +24,6 @@ exports.uuid = function () {
     return '88888888-8888-4888-2888-888888888888'.replace(/[82]/g, (c) => (Math.random() * c * 2 ^ 8).toString(16));
 };
 
-/**
- * Add timestamp to message.
- *
- * @param {*} message Message to timestamp
- * @returns {string} timestamped input message
- */
-function addTimestamp(message) {
-    const date = new Date().toISOString().replace(/\..+$/, '');
-    if (typeof message !== 'string') {
-        message = JSON.stringify(message, null, '    ');
-    }
-    message = '[' + date.replace(/^.+T/, '') + '] ' + message;
-    return message;
-}
 
 /**
  * Log a message to the console
@@ -45,7 +31,7 @@ function addTimestamp(message) {
  * @param {*} message Message to log
  */
 exports.log = function (message) {
-    console.log(addTimestamp(message)); //eslint-disable-line no-console
+    console.log(message); //eslint-disable-line no-console
 };
 
 /**
@@ -54,7 +40,7 @@ exports.log = function (message) {
  * @param {*} message Warning to log
  */
 exports.warn = function (message) {
-    console.warn(addTimestamp(message)); //eslint-disable-line no-console
+    console.warn(message); //eslint-disable-line no-console
 };
 
 /**
@@ -63,7 +49,7 @@ exports.warn = function (message) {
  * @param {*} message Error to log
  */
 exports.error = function (message) {
-    console.error(addTimestamp(message)); //eslint-disable-line no-console
+    console.error(message); //eslint-disable-line no-console
 };
 
 /**
@@ -181,91 +167,87 @@ exports.mergeObjects = function mergeObjects(mergeArrays, mixin) { //eslint-disa
     return res;
 };
 
-/**
- * Proccess post for ignore contitions
- *
- * @param {externals.posts.CleanedPost} post Post to filter
- * @param {filterCallback} callback Completion Callback
- * @returns {null} No return value
- */
-exports.filterIgnoredOnPost = function filterIgnoredOnPost(post, callback) {
-    const flow = (err, msg) => setTimeout(() => callback(err, msg), 0),
-        ignoredUsers = config.core.ignoreUsers,
-        now = Date.now();
-    if (post.trust_level >= 4) {
-        return flow(null, 'Poster is TL4+');
-    }
-    if (ignoredUsers.indexOf(post.username) >= 0) {
-        return flow('ignore', 'Post Creator Ignored');
-    }
-    if (post.trust_level < 1) {
-        return flow('ignore', 'Poster is TL0');
-    }
-    if (post.trust_level === 1) {
-        if (internals.cooldownTimers[post.username] && now < internals.cooldownTimers[post.username]) {
-            return flow('ignore', 'Poster is TL1 on Cooldown');
-        }
-        internals.cooldownTimers[post.username] = now + config.core.cooldownPeriod;
-    }
-    if (post.primary_group_name === 'bots') {
-        return flow('ignore', 'Poster is a Bot');
-    }
-    return flow(null, 'POST OK');
-};
 
-/**
- * Proccess topic for ignore contitions
- *
- * @param {externals.topics.Topic} topic Topic to filter
- * @param {filterCallback} callback Completion Callback
- * @returns {null} No return value
- */
-exports.filterIgnoredOnTopic = function filterIgnoredOnTopic(topic, callback) {
-    const flow = (err, msg) => setTimeout(() => callback(err, msg), 0),
-        ignoredUsers = config.core.ignoreUsers,
-        ignoredCategory = config.core.ignoreCategories;
-    if (topic.details) {
-        const user = topic.details.created_by.username;
-        if (topic.details.notification_level < 1) {
-            return flow('ignore', 'Topic Was Muted');
-        }
-        if (ignoredUsers.indexOf(user) >= 0) {
-            return flow('ignore', 'Topic Creator Ignored');
-        }
-    }
-    if (ignoredCategory.indexOf(topic.category_id) >= 0) {
-        return flow('ignore', 'Topic Category Ignored');
-    }
-    return flow(null, 'TOPIC OK');
-};
-
-/**
- * Filter post/topic for ignore conditions
- *
- * @param {externals.topics.Topic} topic Topic to filter
- * @param {externals.posts.CleanedPost} post Post to filter
- * @param {completionCallback} callback Completion Callback
- */
-exports.filterIgnored = function filterIgnored(topic, post, callback) {
-    async.parallel([
-        (cb) => {
-            exports.filterIgnoredOnPost(post, cb);
-        }, (cb) => {
-            exports.filterIgnoredOnTopic(topic, cb);
-        }
-    ], (err, reason) => {
-        if (err) {
-            exports.warn('Post #' + post.id + ' Ignored: ' + reason.join(', '));
-        }
-        callback(err);
-    });
-};
 
 
 /* istanbul ignore else */
 if (typeof GLOBAL.describe === 'function') {
     //test is running
-    exports.addTimestamp = addTimestamp;
     exports.mergeInner = mergeInner;
-    exports.internals = internals;
 }
+
+
+
+/**
+ * description
+ *
+ */
+exports.mapGet = function mapGet(obj, key) {
+    return (storage.get(obj) || {})[key];
+};
+
+/**
+ * description
+ *
+ */
+exports.mapSet = function mapSet(obj, key, value) {
+    let values = key;
+    if (value) {
+        values = storage.get(obj) || {};
+        values[key] = value;
+    }
+    storage.set(obj, values);
+};
+
+/**
+ * description
+ *
+ */
+exports.parseJSON = function parseJSON(json) {
+    if (!json) {
+        throw new Error('[[invalid-argument:required]]');
+    }
+    if (typeof json === 'string') {
+        try {
+            json = JSON.parse(json);
+        } catch (err) {
+            throw new Error(`[[invalid-json:${err.message}]]`);
+        }
+    }
+    const type = typeof json;
+    if (type !== 'object') {
+        throw new Error(`[[invalid-argument: expected object but received ${type}]]`);
+    }
+    return json;
+};
+
+/**
+ * description
+ *
+ */
+exports.iterate = function iterate(arr, each) {
+    return new Promise((resolve, reject) => {
+        const next = () => {
+            if (!arr || !arr.length) {
+                return resolve();
+            }
+            const post = arr.shift();
+            return each(post)
+                .then(() => next())
+                .catch(reject);
+        };
+        next();
+    });
+};
+
+/**
+ * description
+ *
+ */
+exports.htmlToRaw = function htmlToRaw(markup) {
+    markup = sanitizeHtml(markup,{
+        allowedTags:['code','blockquote'],
+        exclusiveFilter:(frame)=>frame.tag ==='code' || frame.tag==='blockquote'
+    });
+    return markup;
+};
