+++ providers/nodebb/topic.js
@@ -1,5 +1,5 @@
 'use strict';
-const utils = require('./utils');
+const utils = require('../../lib/utils');
 
 exports.bindTopic = function bindTopic(forum) {
     class Topic {
